requirements.txt
streamlit
chromadb
openai
pypdf
python-dotenv
tiktoken
fastembed